IDEX-2666: Fixed run extension on Windows OS
