+++ plugin-sdk/che-plugin-sdk-runner/src/main/java/org/eclipse/che/runner/sdk/SDKRunnerModule.java
@@ -10,6 +10,7 @@
  *******************************************************************************/
 package org.eclipse.che.runner.sdk;
 
+import org.eclipse.che.api.core.util.SystemInfo;
 import org.eclipse.che.api.runner.internal.Runner;
 import org.eclipse.che.inject.DynaModule;
 import com.google.inject.AbstractModule;
@@ -21,7 +22,12 @@ public class SDKRunnerModule extends AbstractModule {
     @Override
     protected void configure() {
         Multibinder.newSetBinder(binder(), Runner.class).addBinding().to(SDKRunner.class);
-        Multibinder.newSetBinder(binder(), ApplicationServer.class).addBinding().to(TomcatServer.class);
+        Multibinder<ApplicationServer> multibinder = Multibinder.newSetBinder(binder(), ApplicationServer.class);
+        if (SystemInfo.isUnix()) {
+            multibinder.addBinding().to(UnixTomcatServer.class);
+        } else if (SystemInfo.isWindows()) {
+            multibinder.addBinding().to(WindowsTomcatServer.class);
+        } else throw new UnsupportedOperationException();
         bind(UpdateService.class);
     }
 }
