+++ plugin-sdk/che-plugin-sdk-runner/src/main/java/org/eclipse/che/runner/sdk/TomcatServer.java
@@ -11,17 +11,13 @@
 package org.eclipse.che.runner.sdk;
 
 import org.eclipse.che.api.core.notification.EventService;
-import org.eclipse.che.api.core.util.CommandLine;
-import org.eclipse.che.api.core.util.ProcessUtil;
-import org.eclipse.che.api.core.util.StreamPump;
-import org.eclipse.che.api.core.util.SystemInfo;
 import org.eclipse.che.api.project.shared.dto.ProjectDescriptor;
 import org.eclipse.che.api.runner.RunnerException;
 import org.eclipse.che.api.runner.internal.ApplicationLogger;
-import org.eclipse.che.api.runner.internal.ApplicationLogsPublisher;
 import org.eclipse.che.api.runner.internal.ApplicationProcess;
 import org.eclipse.che.commons.lang.IoUtil;
 import org.eclipse.che.commons.lang.ZipUtils;
+
 import com.google.common.io.CharStreams;
 import com.google.inject.Singleton;
 
@@ -30,11 +26,11 @@ import org.slf4j.LoggerFactory;
 
 import javax.inject.Inject;
 import javax.inject.Named;
+import java.io.File;
 import java.io.FileReader;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.zip.ZipFile;
 
@@ -45,7 +41,7 @@ import java.util.zip.ZipFile;
  * @author Eugene Voevodin
  */
 @Singleton
-public class TomcatServer implements ApplicationServer {
+public abstract class TomcatServer implements ApplicationServer {
     public static final  String MEM_SIZE_PARAMETER = "runner.tomcat.memory";
     private static final Logger LOG                = LoggerFactory.getLogger(TomcatServer.class);
     private static final String SERVER_XML         =
@@ -67,7 +63,7 @@ public class TomcatServer implements ApplicationServer {
 
     private final int                        memSize;
     private final ApplicationUpdaterRegistry applicationUpdaterRegistry;
-    private final EventService               eventService;
+    protected final EventService eventService;
 
     @Inject
     public TomcatServer(@Named(MEM_SIZE_PARAMETER) int memSize, ApplicationUpdaterRegistry applicationUpdaterRegistry,
@@ -95,15 +91,13 @@ public class TomcatServer implements ApplicationServer {
                                      final SDKRunnerConfiguration runnerConfiguration,
                                      CodeServer.CodeServerProcess codeServerProcess,
                                      ApplicationProcess.Callback callback) throws RunnerException {
-        Path tomcatPath;
-        final Path webappsPath;
-        final Path apiAppContextPath;
-        try {
-            tomcatPath = Files.createDirectory(workDir.toPath().resolve("tomcat"));
-            ZipUtils.unzip(Utils.getTomcatBinaryDistribution().openStream(), tomcatPath.toFile());
-            webappsPath = tomcatPath.resolve("webapps");
+        final Path tomcatPath = workDir.toPath().resolve("tomcat");
+        final Path webappsPath = tomcatPath.resolve("webapps");
+        final Path apiAppContextPath = webappsPath.resolve("che");
 
-            apiAppContextPath = webappsPath.resolve("che");
+        try {
+            Files.createDirectory(tomcatPath);
+            ZipUtils.unzip(Utils.getTomcatBinaryDistribution().openStream(), tomcatPath.toFile());
             ZipUtils.unzip(new java.io.File(warToDeploy.getName()), apiAppContextPath.toFile());
             generateServerXml(tomcatPath.toFile(), runnerConfiguration);
             //add JAR with extension to 'api' application's 'lib' directory
@@ -112,12 +106,7 @@ public class TomcatServer implements ApplicationServer {
             throw new RunnerException(e);
         }
 
-        ApplicationProcess process;
-        if (SystemInfo.isUnix()) {
-            process = startUnix(workDir, runnerConfiguration, codeServerProcess, callback);
-        } else {
-            process = startWindows(workDir, runnerConfiguration, codeServerProcess, callback);
-        }
+        ApplicationProcess process = start(workDir, runnerConfiguration, codeServerProcess, callback);
 
         // TODO: unregister updater
         registerUpdater(process, new ApplicationUpdater() {
@@ -144,7 +133,7 @@ public class TomcatServer implements ApplicationServer {
         return process;
     }
 
-    private void registerUpdater(ApplicationProcess process, ApplicationUpdater updater) {
+    protected void registerUpdater(ApplicationProcess process, ApplicationUpdater updater) {
         applicationUpdaterRegistry.registerUpdater(process, updater);
     }
 
@@ -163,238 +152,55 @@ public class TomcatServer implements ApplicationServer {
         return "Tomcat Server";
     }
 
-    // *nix
+    /**
+     * Start application process
+     * @param appDir application directory
+     * @param runnerCfg runner configuration
+     * @param codeServerProcess code server process
+     * @param callback some actions after start application process
+     * @return launched application process
+     */
+    protected abstract ApplicationProcess start(java.io.File appDir,
+                                      SDKRunnerConfiguration runnerCfg,
+                                      CodeServer.CodeServerProcess codeServerProcess,
+                                      ApplicationProcess.Callback callback) throws RunnerException;
 
-    protected ApplicationProcess startUnix(java.io.File appDir, SDKRunnerConfiguration runnerCfg,
-                                           CodeServer.CodeServerProcess codeServerProcess, ApplicationProcess.Callback callback)
-            throws RunnerException {
-        final java.io.File startUpScriptFile;
-        final java.io.File logsDir = new java.io.File(appDir, "logs");
-        try {
-            startUpScriptFile = genStartUpScriptUnix(appDir, runnerCfg);
-            updateSetenvFileUnix(appDir, runnerCfg);
-            Files.createDirectory(logsDir.toPath());
-        } catch (IOException e) {
-            throw new RunnerException(e);
-        }
-        final List<java.io.File> logFiles = new ArrayList<>(1);
-        logFiles.add(new java.io.File(logsDir, "output.log"));
-        return new TomcatProcess(appDir, startUpScriptFile, logFiles, runnerCfg, callback, codeServerProcess, eventService);
-    }
-
-    private java.io.File genStartUpScriptUnix(java.io.File appDir, SDKRunnerConfiguration runnerConfiguration) throws IOException {
-        final String startupScript = "#!/bin/sh\n" +
-                                     exportEnvVariablesUnix(runnerConfiguration) +
-                                     "cd tomcat\n" +
-                                     "chmod +x bin/*.sh\n" +
-                                     catalinaUnix(runnerConfiguration) +
-                                     "PID=$!\n" +
-                                     "echo \"$PID\" > ../run.pid\n" +
-                                     "wait $PID";
-        final java.io.File startUpScriptFile = new java.io.File(appDir, "startup.sh");
-        Files.write(startUpScriptFile.toPath(), startupScript.getBytes());
-        if (!startUpScriptFile.setExecutable(true, false)) {
-            throw new IOException("Unable to update attributes of the startup script");
-        }
-        return startUpScriptFile;
-    }
-
-    private void updateSetenvFileUnix(java.io.File tomcatDir, SDKRunnerConfiguration runnerCfg) throws IOException {
-        final Path setenvShPath = tomcatDir.toPath().resolve("tomcat/bin/setenv.sh");
-        final String setenvShContent =
-                new String(Files.readAllBytes(setenvShPath)).replace("${PORT}", Integer.toString(runnerCfg.getHttpPort()));
-        Files.write(setenvShPath, setenvShContent.getBytes());
-    }
-
-    private String exportEnvVariablesUnix(SDKRunnerConfiguration runnerConfiguration) {
-        int memory = runnerConfiguration.getMemory();
-        if (memory <= 0) {
-            memory = getMemSize();
-        }
-        final StringBuilder export = new StringBuilder();
-        export.append(String.format("export CATALINA_OPTS=\"-Xms%dm -Xmx%dm\"%n", memory, memory));
-        export.append(String.format("export SERVER_PORT=%d%n", runnerConfiguration.getHttpPort()));
-        final int debugPort = runnerConfiguration.getDebugPort();
-        if (debugPort > 0) {
-            /*
-            From catalina.sh:
-            -agentlib:jdwp=transport=$JPDA_TRANSPORT,address=$JPDA_ADDRESS,server=y,suspend=$JPDA_SUSPEND
-             */
-            export.append(String.format("export JPDA_ADDRESS=%d%n", debugPort));
-            export.append(String.format("export JPDA_TRANSPORT=%s%n", "dt_socket"));
-            export.append(String.format("export JPDA_SUSPEND=%s%n", runnerConfiguration.isDebugSuspend() ? "y" : "n"));
-        }
-        return export.toString();
-    }
-
-    private String catalinaUnix(SDKRunnerConfiguration runnerConfiguration) {
-        final boolean debug = runnerConfiguration.getDebugPort() > 0;
-        if (debug) {
-            return "./bin/catalina.sh jpda run 2>&1 | tee ../logs/output.log &\n";
-        }
-        return "./bin/catalina.sh run 2>&1 | tee ../logs/output.log &\n";
-    }
-
-    // Windows
-
-    protected ApplicationProcess startWindows(java.io.File appDir, SDKRunnerConfiguration runnerConfiguration,
-                                              CodeServer.CodeServerProcess codeServerProcess, ApplicationProcess.Callback callback) {
-        throw new UnsupportedOperationException();
-    }
-
-    private static class TomcatProcess extends ApplicationProcess {
-        final int                          httpPort;
-        final List<java.io.File>           logFiles;
-        final int                          debugPort;
-        final java.io.File                 startUpScriptFile;
-        final java.io.File                 workDir;
+    protected static class TomcatLogger implements ApplicationLogger {
+        final List<File>                   logFiles;
         final CodeServer.CodeServerProcess codeServerProcess;
-        final Callback                     callback;
-        final String                       workspace;
-        final String                       project;
-        final long                         id;
-        final EventService                 eventService;
 
-        ApplicationLogger logger;
-        Process           process;
-        StreamPump        output;
-
-        TomcatProcess(java.io.File appDir, java.io.File startUpScriptFile, List<java.io.File> logFiles,
-                      SDKRunnerConfiguration runnerCfg, Callback callback, CodeServer.CodeServerProcess codeServerProcess,
-                      EventService eventService) {
-            this.httpPort = runnerCfg.getHttpPort();
+        TomcatLogger(List<java.io.File> logFiles, CodeServer.CodeServerProcess codeServerProcess) {
             this.logFiles = logFiles;
-            this.debugPort = runnerCfg.getDebugPort();
-            this.startUpScriptFile = startUpScriptFile;
             this.codeServerProcess = codeServerProcess;
-            this.workDir = appDir;
-            this.callback = callback;
-            this.eventService = eventService;
-            this.workspace = runnerCfg.getRequest().getWorkspace();
-            this.project = runnerCfg.getRequest().getProject();
-            this.id = runnerCfg.getRequest().getId();
         }
 
         @Override
-        public synchronized void start() throws RunnerException {
-            if (process != null && ProcessUtil.isAlive(process)) {
-                throw new IllegalStateException("Process is already started");
+        public void getLogs(Appendable output) throws IOException {
+            for (java.io.File logFile : logFiles) {
+                output.append(String.format("%n====> %1$s <====%n%n", logFile.getName()));
+                try (FileReader r = new FileReader(logFile)) {
+                    CharStreams.copy(r, output);
+                }
+                output.append(System.lineSeparator());
             }
             try {
-                process = Runtime.getRuntime().exec(new CommandLine(startUpScriptFile.getAbsolutePath()).toShellCommand(), null, workDir);
-                logger = new ApplicationLogsPublisher(new TomcatLogger(logFiles, codeServerProcess), eventService, id, workspace, project);
-                output = new StreamPump();
-                output.start(process, logger);
-                try {
-                    codeServerProcess.start();
-                } catch (Exception e) {
-                    LOG.error(e.getMessage(), e);
-                }
-                LOG.debug("Start Tomcat at port {}, application {}", httpPort, workDir);
-            } catch (IOException e) {
-                throw new RunnerException(e);
+                codeServerProcess.getLogs(output);
+            } catch (Exception ignore) {
             }
         }
 
         @Override
-        public synchronized void stop() throws RunnerException {
-            if (process == null) {
-                throw new IllegalStateException("Process is not started yet");
-            }
-            // Use ProcessUtil.kill(process) because java.lang.Process.destroy() method doesn't
-            // kill all child processes (see http://bugs.sun.com/view_bug.do?bug_id=4770092).
-            ProcessUtil.kill(process);
-            if (output != null) {
-                output.stop();
-            }
-            try {
-                codeServerProcess.stop();
-            } catch (Exception e) {
-                LOG.error(e.getMessage(), e);
-            }
-            callback.stopped();
-            LOG.debug("Stop Tomcat at port {}, application {}", httpPort, workDir);
+        public String getContentType() {
+            return "text/plain";
         }
 
         @Override
-        public int waitFor() throws RunnerException {
-            synchronized (this) {
-                if (process == null) {
-                    throw new IllegalStateException("Process is not started yet");
-                }
-            }
-            try {
-                process.waitFor();
-            } catch (InterruptedException e) {
-                Thread.interrupted();
-                ProcessUtil.kill(process);
-            } finally {
-                if (output != null) {
-                    output.stop();
-                }
-            }
-            return process.exitValue();
+        public void writeLine(String line) throws IOException {
+            // noop since logs already redirected to the file
         }
 
         @Override
-        public synchronized int exitCode() throws RunnerException {
-            if (process == null || ProcessUtil.isAlive(process)) {
-                return -1;
-            }
-            return process.exitValue();
-        }
-
-        @Override
-        public synchronized boolean isRunning() throws RunnerException {
-            return process != null && ProcessUtil.isAlive(process);
-        }
-
-        @Override
-        public synchronized ApplicationLogger getLogger() throws RunnerException {
-            if (logger == null) {
-                // is not started yet
-                return ApplicationLogger.DUMMY;
-            }
-            return logger;
-        }
-
-        private static class TomcatLogger implements ApplicationLogger {
-            final List<java.io.File>           logFiles;
-            final CodeServer.CodeServerProcess codeServerProcess;
-
-            TomcatLogger(List<java.io.File> logFiles, CodeServer.CodeServerProcess codeServerProcess) {
-                this.logFiles = logFiles;
-                this.codeServerProcess = codeServerProcess;
-            }
-
-            @Override
-            public void getLogs(Appendable output) throws IOException {
-                for (java.io.File logFile : logFiles) {
-                    output.append(String.format("%n====> %1$s <====%n%n", logFile.getName()));
-                    try (FileReader r = new FileReader(logFile)) {
-                        CharStreams.copy(r, output);
-                    }
-                    output.append(System.lineSeparator());
-                }
-                try {
-                    codeServerProcess.getLogs(output);
-                } catch (Exception ignore) {
-                }
-            }
-
-            @Override
-            public String getContentType() {
-                return "text/plain";
-            }
-
-            @Override
-            public void writeLine(String line) throws IOException {
-                // noop since logs already redirected to the file
-            }
-
-            @Override
-            public void close() throws IOException {
-            }
+        public void close() throws IOException {
         }
     }
 }
